[Java] Extract common base class for MutableDirectBuffer tests.
